eatgo-admin-api/src/test/java/kr/co/fastcampus/eatgo/application/UserServiceTests.java
package kr.co.fastcampus.eatgo.application;

import kr.co.fastcampus.eatgo.domain.User;
import kr.co.fastcampus.eatgo.domain.UserRepository;
import org.junit.Before;
import org.junit.Test;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;

import java.util.List;

import static org.hamcrest.core.Is.is;
import static org.junit.Assert.assertThat;
import static org.mockito.ArgumentMatchers.any;
import static org.mockito.BDDMockito.given;

public class UserServiceTests {

    private UserService userService;

    @Mock
    private UserRepository userRepository;

    @Before
    public void setUp(){
        MockitoAnnotations.initMocks(this);
        userService = new UserService(userRepository);
    }

    @Test
    public void getUsers(){

        List<User> mockUsers = userService.getUsers();
        mockUsers.add(User.builder()
                .email("tester@example.com")
                .name("tester")
                .level(3L)
                .build());

        given(userRepository.findAll()).willReturn(mockUsers);

        User user = mockUsers.get(0);
        assertThat(user.getName(), is("tester"));
    }

    @Test
    public void addUser(){
        String email ="admin@exmaple.com";
        String name ="Administrator";

        User mockUser = User.builder()
                .email(email)
                .name(name)
                .build();

        given(userRepository.save(any())).willReturn(mockUser);

        User user = userService.addUser(email, name);

        assertThat(user.getName(), is(name));
    }
}